bake in spicey
