+++ tests/features/spice-analysis/spice-analysis01-platform-config.test.tsx
@@ -1,49 +1,8 @@
 import { test, expect } from "bun:test"
 import { getTestFixture } from "tests/fixtures/get-test-fixture"
-import { simulate, spiceyTranToVGraphs } from "spicey"
-import type { SimulationExperiment } from "circuit-json"
 
 test("spice-analysis01-platform-config", async () => {
   const { circuit } = getTestFixture()
-
-  circuit.platform = {
-    spiceEngineMap: {
-      spicey: {
-        async simulate(spiceString: string) {
-          const simulation_experiment_id = "spice-experiment-1"
-
-          // Add .tran directive before .END if not present
-          let spiceNetlist = spiceString
-          if (!spiceNetlist.includes(".tran")) {
-            spiceNetlist = spiceNetlist.replace(
-              /\.END/i,
-              ".tran 1us 10ms\n.END",
-            )
-          }
-
-          // Run spicey simulation
-          const { circuit: parsedCircuit, tran } = simulate(spiceNetlist)
-
-          // Convert transient results to voltage graphs
-          const voltageGraphs = spiceyTranToVGraphs(
-            tran,
-            parsedCircuit,
-            simulation_experiment_id,
-          )
-
-          return {
-            simulationResultCircuitJson: [
-              {
-                type: "simulation_experiment",
-                simulation_experiment_id,
-              } as SimulationExperiment,
-              ...voltageGraphs,
-            ],
-          }
-        },
-      },
-    },
-  }
 
   circuit.add(
     <board schMaxTraceDistance={10}>
